+++ NestEngine/src/Nest/graphics/IndexBuffer.h
@@ -6,18 +6,18 @@ namespace Nest
 {
 	class IndexBuffer
 	{
-		unsigned int m_count = 0;
+		size_t m_count = 0;
 		unsigned int m_rendererID = 0;
 
 	public:
 		IndexBuffer() : m_count(0), m_rendererID(0) {}
-		IndexBuffer(unsigned int count, unsigned int *indices);
+		IndexBuffer(size_t count, unsigned int *indices);
 		IndexBuffer(const IndexBuffer&) = delete;
 		~IndexBuffer();
 
 		void bind() const;
 		void unbind() const;
 
-		unsigned int getCount() const { return m_count; }
+		size_t getCount() const { return m_count; }
 	};
 }
